+++ postcss.config.js
@@ -1,0 +1,5 @@
+/* eslint-disable no-undef */
+
+module.exports = {
+  plugins: ['autoprefixer'],
+}